+++ R/instances.R
@@ -15,8 +15,8 @@
 
 instances <- function(term, instances = 1){
     term <- match.call()$term
-    if (!"inst" %in% names(formals(match.fun(term[[1]]))))
-        stop(term[[1]], " has no inst argumnt")
+    if (!"inst" %in% names(formals(match.fun(eval(term[[1]])))))
+        stop(deparse(term[[1]]), " has no inst argument")
     termList <- vector(mode = "list", length = instances)
     for (i in seq(instances)) {
         termList[[i]] <- term