+++ tests/testthat/test-doubleUnidiff.R
@@ -9,6 +9,7 @@ test_that("double unidiff model as expected for cautres data", {
                              Mult(Exp(election), religion:vote) +
                              Mult(Exp(election), class:vote),
                          family = poisson, data = cautres, verbose = FALSE)
+    doubleUnidiff$family$dispersion <- 1
     expect_equal(round(deviance(doubleUnidiff), 2), 133.04)
     expect_known_value(doubleUnidiff,
                        file = test_path("outputs/doubleUnidiff.rds"))